Updated a few dependencies
Added a "proxy.conf.json" and a proxy option in angular.json to redirect the MagicSeaWeed API all to this proxy and bypass the CORS security on the msw server.

Initiated createSpot and UpdateSpot comonents using the same form. ---> Added the ngModel to the Form (the previous highlight directive does not work any more now).

Modified the httpClient.Get() call to a jsonP() call for the REST call to the msw API (CORS policy at msw server required it)

Updated the SurfService component

All the REST calls to "TheWatchTowerMachnie" now work fine.